fix(auth): broken export config
