+++ src/components/Navbar/index.tsx
@@ -1,9 +1,9 @@
 import { use } from "react";
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 import Profile from "../User/Profile";
 import SignIn from "../User/SignIn";
 import SearchBar from "./SearchBar";
+import { authOptions } from "@/app/api/auth/[...nextauth]/authOptions";
 
 async function getSession() {
   const session = await getServerSession(authOptions);
